+++ lib/Conversion/TorchToTosa/CMakeLists.txt
@@ -12,8 +12,11 @@ add_mlir_conversion_library(TorchMLIRTorchToTosa
   LINK_LIBS PUBLIC
   MLIRIR
   MLIRPass
+  MLIRRewrite
   MLIRTosaDialect
+  MLIRTransformUtils
   TorchMLIRConversionUtils
+  TorchMLIRTorchConversionPasses
   TorchMLIRTorchDialect
 )
