+++ lib/Dialect/TorchConversion/IR/CMakeLists.txt
@@ -13,6 +13,8 @@ add_mlir_dialect_library(TorchMLIRTorchConversionDialect
   MLIRIR
   MLIRSupport
   MLIRSideEffectInterfaces
+  MLIRArithDialect
+  TorchMLIRTorchDialect
 )
 
 torch_mlir_target_includes(TorchMLIRTorchConversionDialect)
